requirements.txt
autopep8==2.3.2
boto3
blinker==1.9.0
click==8.2.1
python-dotenv
psycopg2-binary
cmdstanpy==1.2.5
contourpy==1.3.3
cycler==0.12.1
Flask==3.0.3
fonttools==4.59.1
gunicorn==22.0.0
holidays==0.79
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.5
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
prophet==1.1.7
pycodestyle==2.14.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
stanio==0.5.1
tqdm==4.67.1
tzdata==2025.2
Werkzeug==3.0.6
